css/css-values/minmax-number-serialize.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-values-4/#comp-func">
<link rel="help" href="https://drafts.csswg.org/css-values-4/#numbers">
<link rel="help" href="https://drafts.csswg.org/css-values-4/#calc-serialize">
<link rel="author" title="Xiaocheng Hu" href="mailto:xiaochengh@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../support/numeric-testcommon.js"></script>
<div id=target></div>
<script>
const prop = 'opacity';
const base = '.123';
const el = document.querySelector("#target");
function serializeSpecified(t, e) {
	test(()=>{
		el.style[prop] = base;
		el.style[prop] = t;
		const tValue = el.style[prop];
		assert_not_equals(tValue, base, `'${t}' should be valid in ${prop}.`);

		el.style[prop] = base;
		el.style[prop] = e;
		const eValue = el.style[prop];
		assert_not_equals(eValue, base, `'${e}' should be valid in ${prop}.`);
		assert_equals(e, eValue, `'${e}' should round-trip exactly in specified values.`);

		assert_equals(tValue, e, `'${t}' and '${e}' should serialize the same in specified values.`);
	});
}
function serializeComputed(t, e) {
	test(()=>{
		el.style[prop] = base;
		el.style[prop] = t;
		const tValue = getComputedStyle(el)[prop];
		assert_not_equals(tValue, base, `'${t}' should be valid in ${prop}.`);

		el.style[prop] = base;
		el.style[prop] = e;
		const eValue = getComputedStyle(el)[prop];
		assert_not_equals(eValue, base, `'${e}' should be valid in ${prop}.`);
		assert_equals(e, eValue, `'${e}' should round-trip exactly in computed values.`);

		assert_equals(tValue, e, `'${t}' and '${e}' should serialize the same in computed values.`);
	});
}

test_math_specified('min(1)', 'calc(1)', {type:"number"});
test_math_computed('min(1)', 'calc(1)', {type:"number"});
test_math_used('min(1)', '1', {type:"number"});
test_math_specified('max(1)', 'calc(1)', {type:"number"});
test_math_computed('max(1)', 'calc(1)', {type:"number"});
test_math_used('max(1)', '1', {type:"number"});

test_math_specified('min(1, 2, 3)', 'calc(1)', {type:"number"});
test_math_computed('min(1, 2, 3)', 'calc(1)', {type:"number"});
test_math_used('min(1, 2, 3)', '1', {type:"number"});
test_math_specified('max(1, 2, 3)', 'calc(3)', {type:"number"});
test_math_computed('max(1, 2, 3)', 'calc(3)', {type:"number"});
test_math_used('max(1, 2, 3)', '3', {type:"number"});

test_math_specified('min(3, 2, 1)', 'calc(1)', {type:"number"});
test_math_computed('min(3, 2, 1)', 'calc(1)', {type:"number"});
test_math_used('min(3, 2, 1)', '1', {type:"number"});
test_math_specified('max(3, 2, 1)', 'calc(3)', {type:"number"});
test_math_computed('max(3, 2, 1)', 'calc(3)', {type:"number"});
test_math_used('max(3, 2, 1)', '3', {type:"number"});

test_math_specified('calc(min(1) + min(2))', 'calc(3)', {type:"number"});
test_math_used('calc(min(1) + min(2))', '3', {type:"number"});
test_math_specified('calc(max(1) + max(2))', 'calc(3)', {type:"number"});
test_math_used('calc(max(1) + max(2))', '3', {type:"number"});
test_math_specified('calc(1 + min(1))', 'calc(2)', {type:"number"});
test_math_used('calc(1 + min(1))', '2', {type:"number"});
test_math_specified('calc(max(1) + 1)', 'calc(2)', {type:"number"});
test_math_used('calc(max(1) + 1)', '2', {type:"number"});
</script>
